backend/requirements.txt
asgiref==3.2.10
astroid==2.4.2
beautifulsoup4==4.9.1
boto==2.49.0
boto3==1.15.16
botocore==1.18.16
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
Django==3.1.2
django-bootstrap4==2.2.0
django-cors-headers==3.4.0
django-filter==2.3.0
django-pandas==0.6.2
django-storages==1.10.1
django-timestampedmodel==0.1.0
djangorestframework==3.11.0
google-api-core==1.21.0
google-api-python-client==1.10.0
google-auth==1.19.1
google-auth-httplib2==0.0.4
googleapis-common-protos==1.52.0
httplib2==0.18.1
idna==2.10
importlib-metadata==1.7.0
isort==4.3.21
jmespath==0.10.0
lazy-object-proxy==1.4.3
mailchimp-marketing==3.0.22
mailchimp3==3.0.14
Markdown==3.2.2
mccabe==0.6.1
numpy==1.19.2
oauth2client==4.1.3
pandas==0.24.0
protobuf==3.12.2
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2020.1
requests==2.24.0
rsa==4.6
s3transfer==0.3.3
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
typed-ast==1.4.1
uritemplate==3.0.1
urllib3==1.25.9
wrapt==1.12.1
zipp==3.1.0
